Factor out vacancy filling into a separate service
