+++ app/helpers/event_signup_mailer_helper.rb
@@ -12,7 +12,7 @@ module EventSignupMailerHelper
   end
 
   def move_result_description(move_result)
-    move_result = EventWithdrawService::SignupMoveResult.from_h(move_result) if move_result.is_a?(Hash)
+    move_result = EventVacancyFillService::SignupMoveResult.from_h(move_result) if move_result.is_a?(Hash)
     signup = move_result.signup
 
     "#{signup.user_con_profile.name}: #{prev_signup_state_description(move_result)} → #{new_signup_state_description(move_result)}"
